Parse and print more LUA syntax.
I changed the AST to fit closer to LUA's true syntax. There are some
glitches, though:

    * Statements can contain PreExp that are not function calls;
    * assignments can assign non lvalues.

Hence, what is parsed is no longer a subset of LUA's syntax, it is no
longer comparable by inclusion.

Now, we use the Text.Parsec.Language feature to define base parsers, it
permits not to say that "end" is a variable, which is problematic with
the current syntax: if "end" is recognized as a variable, there is no
way to close a "do" block (for instance).