Add full support of local bindings.
Local bindings must be a special statement, I added the Bind constructor
to reify them. I also had to add BindFun because of LUA's semantics. We
cannot replace this chunk of code:

    local function x(bar)
        return 42;
    end

by this one:

    local x = function (bar)
	return 42;
    end;

Since it is syntactic sugar for:

    local x;
    x = function (bar)
	return 42;
    end;

That is, *two* statements, hence the need of the BindFun special
statement (because the lstat parser must read one statement from the
string and return *one* statement). It makes a difference when dealing
with recursive functions.  Global functions do not have this problem and
I translate them directly to their assignment form.